Fixed solution to second problem of day 20.
Luckily, the first attempt didn't fail: the stopping criterion is more complicated than
just looping 500 times.

The "fix" is actually also not sufficient but it is better than the previous attempt.

One'd have to check if the remaining lines did intersect and one'd have to check for every
to particles that they either had different velocities, or that they had the same velocity
but passed each other already, etc.